Added: IP consumer (static and dynamic). Need test. NDN Provider (dynamic). Tested with NDN consumer. NDN consumer (static and dynamic). Static need test.
Modified the time calculation: 
For all the time that we calculate, we round it to the next second if it is not on the boundry of a second.
E.g., 1.253 -> 2.000, 1.000 -> 1.000
Reason: In HTTP, the time is in seconds, but in NDN and file, the time is in ms.

TODO: need to add main files for different purposes.
Later on, people can start some programs with:
java -classpath ***.jar edu.rutgers.winlab.** (main class name)